LLMAdapter/requirements.txt
anthropic>=0.10.0
fastapi>=0.103.0
uvicorn>=0.23.2
websockets>=11.0.3
python-dotenv>=1.0.0
sse-starlette>=1.6.5